# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "AccessKeyResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last4": {
            "title": "Last4",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "scopes": {
            "items": {
              "const": "search",
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "last4",
          "scopes",
          "created_at",
          "last_used_at",
          "expires_at",
          "revoked_at"
        ],
        "title": "AccessKeyResponse",
        "type": "object"
      },
...
      "CreateAccessKeyRequest": {
        "additionalProperties": false,
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "scopes": {
            "items": {
              "const": "search",
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": ["name"],
        "title": "CreateAccessKeyRequest",
        "type": "object"
      },
      "CreatedAccessKeyResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "last4": {
            "title": "Last4",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "scopes": {
            "items": {
              "const": "search",
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "prefix",
          "last4",
          "scopes",
          "created_at",
          "last_used_at",
          "expires_at",
          "revoked_at",
          "key"
        ],
        "title": "CreatedAccessKeyResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicSearchFilters": {
        "additionalProperties": false,
        "description": "Filters supported by the public HTTP search contract.",
        "properties": {
          "authors": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Authors",
            "type": "array"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Categories",
            "type": "array"
          },
          "date_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date From"
          },
          "date_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date To"
          }
        },
        "title": "PublicSearchFilters",
        "type": "object"
      },
...
      "PublicSearchHit": {
        "additionalProperties": false,
        "description": "A ranked paper result without internal search diagnostics.",
        "properties": {
          "abstract": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstract"
          },
          "arxiv_id": {
            "minLength": 1,
            "title": "Arxiv Id",
            "type": "string"
          },
          "arxiv_url": {
            "format": "uri",
            "minLength": 1,
            "title": "Arxiv Url",
            "type": "string"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "pdf_url": {
            "format": "uri",
            "minLength": 1,
            "title": "Pdf Url",
            "type": "string"
          },
          "rank": {
            "description": "Authoritative order within this response.",
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "description": "Unnormalized retrieval signal; compare only within the current response, never across queries, strengths, indexes, models, or time.",
            "title": "Score",
            "type": "number"
          },
          "submitted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "rank",
          "score",
          "arxiv_id",
          "title",
          "authors",
          "abstract",
          "categories",
          "submitted_at",
          "updated_at",
          "version",
          "arxiv_url",
          "pdf_url"
        ],
        "title": "PublicSearchHit",
        "type": "object"
      },
      "PublicSearchRequest": {
        "additionalProperties": false,
        "description": "Stable public search input, isolated from internal pipeline controls.",
        "examples": [
          {
            "filters": {
              "authors": [],
              "categories": ["cs.AI", "cs.IR"],
              "date_from": "2020-01-01"
            },
            "limit": 10,
            "query": "retrieval augmented generation",
            "strength": "standard"
          }
        ],
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/PublicSearchFilters"
          },
          "limit": {
            "default": 10,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "strength": {
            "$ref": "#/components/schemas/SearchStrength",
            "default": "standard"
          }
        },
        "required": ["query"],
        "title": "PublicSearchRequest",
        "type": "object"
      },
      "PublicSearchResponse": {
        "additionalProperties": false,
        "description": "Public search response; rank is authoritative and score is unnormalized.",
        "examples": [
          {
            "degraded": true,
            "elapsed_ms": 842.37,
            "hits": [
              {
                "arxiv_id": "2401.12345",
                "arxiv_url": "https://arxiv.org/abs/2401.12345",
                "authors": ["Example Author"],
                "categories": ["cs.AI", "cs.IR"],
                "pdf_url": "https://arxiv.org/pdf/2401.12345",
                "rank": 1,
                "score": 12.75,
                "submitted_at": "2024-01-20T00:00:00Z",
                "title": "A Paper About Retrieval",
                "updated_at": "2024-03-05T00:00:00Z",
                "version": 2
              }
            ],
            "query": "retrieval augmented generation",
            "result_count": 1,
            "strength": "standard"
          }
        ],
        "properties": {
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "elapsed_ms": {
            "minimum": 0.0,
            "title": "Elapsed Ms",
            "type": "number"
          },
          "hits": {
            "items": {
              "$ref": "#/components/schemas/PublicSearchHit"
            },
            "title": "Hits",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "result_count": {
            "minimum": 0.0,
            "title": "Result Count",
            "type": "integer"
          },
          "strength": {
            "$ref": "#/components/schemas/SearchStrength"
          }
        },
        "required": ["query", "strength", "degraded", "hits", "result_count", "elapsed_ms"],
        "title": "PublicSearchResponse",
        "type": "object"
      },
...
      "SearchStrength": {
        "description": "Public search quality presets.",
        "enum": ["standard", "thorough"],
        "title": "SearchStrength",
        "type": "string"
      },
      "SessionResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": ["id", "created_at", "last_seen_at", "expires_at", "current"],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered academic paper search engine",
    "title": "Scholight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/sessions": {
      "get": {
        "operationId": "get_sessions_auth_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionResponse"
                  },
                  "title": "Response Get Sessions Auth Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Sessions",
        "tags": ["sessions"]
      }
    },
...
    "/search": {
      "post": {
        "operationId": "search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {}
        ],
        "summary": "Search",
        "tags": ["search"]
      }
    },
...
    "/user/access-keys": {
      "get": {
        "operationId": "get_keys_user_access_keys_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccessKeyResponse"
                  },
                  "title": "Response Get Keys User Access Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get Keys",
        "tags": ["access-keys"]
      },
      "post": {
        "operationId": "create_key_user_access_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccessKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedAccessKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Key",
        "tags": ["access-keys"]
      }
    },
...
  }
}